move and rename files to unify project structure
