+++ src/components/Footer/footer.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./footer.css";
+import "./Footer.css";
 
 function Footer({ children }) {
   return (
