+++ src/components/User/User.js
@@ -1,4 +1,4 @@
-import React, {useContext, useEffect} from "react";
+import React, { useContext } from "react";
 import { UserContext} from "../../contexts/UserContext";
 import { Button } from 'react-bootstrap';
 import "./User.css"
